+++ MatrixSDK/Crypto/Data/MXCryptoConstants.h
@@ -33,6 +33,11 @@ FOUNDATION_EXPORT NSString *const kMXCryptoMegolmAlgorithm;
  */
 FOUNDATION_EXPORT NSString *const kMXCryptoMegolmBackupAlgorithm;
 
+/**
+ MXKeyProvider identifier for a 32 bytes long key to pickle secrets managed by the olm library.
+ */
+FOUNDATION_EXPORT NSString *const MXCryptoOlmPickleKeyDataType;
+
 
 #pragma mark - Encrypting error